pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cfi-ai"
version = "0.2.0"
description = "Terminal-first agentic assistant"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.67",
    "rich>=14.0",
    "prompt-toolkit>=3.0",
]

[project.scripts]
cfi-ai = "cfi_ai.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/cfi_ai"]

[tool.pytest.ini_options]
testpaths = ["tests"]
